Added README
"The Economy, Stupid!" is a macroeconomics simulation game that Insight Machines created as its very first project, more than 10 years ago. It was a venture into the world of macroeconomic simulations and serious games. The game runs on Java so you'll need a Java Runtime to run it. You can get the newest version from [here](https://java.com/en/download/). 

The repository contains both the game's source code as well as the compiled version - just for convenience. To run the game, go to the "dist" directory and in the command line type:
```java -jar theeconomystupid-1.0.0.jar```

The game was translated to Spanish by Hugo Sanchez, the owner of ["Simulación Política"](http://simulacionpolitica.blogspot.com/) - a blog devoted to games about political strategy and economics.